music service update

+++ MusicService/src/main/java/com/artillery/musicservice/data/MusicLocalUtils.java
@@ -19,7 +19,7 @@ public class MusicLocalUtils {
     private static String name;
     private static String artist;
     private static String path;
-    private static long duration;
+    private static int duration;
     private static int size;
     private static String album;
     private static int id;
@@ -48,7 +48,7 @@ public class MusicLocalUtils {
                 id = cursor.getInt(cursor.getColumnIndexOrThrow(MediaStore.Audio.Media._ID));
                 artist = cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Audio.Media.ARTIST));
                 path = cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Audio.Media.DATA));
-                duration = cursor.getLong(cursor.getColumnIndexOrThrow(MediaStore.Audio.Media.DURATION));
+                duration = cursor.getInt(cursor.getColumnIndexOrThrow(MediaStore.Audio.Media.DURATION));
                 size = cursor.getInt(cursor.getColumnIndexOrThrow(MediaStore.Audio.Media.SIZE));
                 album = cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Audio.Media.ALBUM));
                 //把歌曲名字和歌手切割开
